+++ src/dash-table/components/ControlledTable/index.tsx
@@ -21,6 +21,7 @@ import dropdownHelper from 'dash-table/components/dropdownHelper';
 import derivedTable from 'dash-table/derived/table';
 import derivedTableFragments from 'dash-table/derived/table/fragments';
 import isEditable from 'dash-table/derived/cell/isEditable';
+import { derivedTableStyle } from 'dash-table/derived/style';
 
 const sortNumerical = R.sort<number>((a, b) => a - b);
 
@@ -31,6 +32,7 @@ interface IState {
 export default class ControlledTable extends PureComponent<ControlledTableProps, IState> {
     private readonly stylesheet: Stylesheet;
     private readonly tableFn: () => JSX.Element[][];
+    private readonly tableStyle = derivedTableStyle();
 
     constructor(props: ControlledTableProps) {
         super(props);
@@ -53,11 +55,11 @@ export default class ControlledTable extends PureComponent<ControlledTableProps,
     }
 
     private updateStylesheet() {
-        const { table_style } = this.props;
+        const { css } = this.props;
 
         R.forEach(({ selector, rule }) => {
             this.stylesheet.setRule(selector, rule);
-        }, table_style);
+        }, css);
     }
 
     componentDidMount() {
@@ -620,7 +622,8 @@ export default class ControlledTable extends PureComponent<ControlledTableProps,
             id,
             content_style,
             n_fixed_columns,
-            n_fixed_rows
+            n_fixed_rows,
+            style
         } = this.props;
 
         const classes = [
@@ -653,14 +656,16 @@ export default class ControlledTable extends PureComponent<ControlledTableProps,
         const rawTable = this.tableFn();
         const grid = derivedTableFragments(n_fixed_columns, n_fixed_rows, rawTable);
 
+        const tableStyle = this.tableStyle(style).style;
+
         return (<div
             id={id}
             onCopy={this.onCopy}
             onKeyDown={this.handleKeyDown}
             onPaste={this.onPaste}
         >
-            <div className={containerClasses.join(' ')}>
-                <div className={classes.join(' ')}>
+            <div className={containerClasses.join(' ')} style={tableStyle}>
+                <div className={classes.join(' ')} style={tableStyle}>
                     {grid.map((row, rowIndex) => (<div
                         key={`r${rowIndex}`}
                         ref={`r${rowIndex}`}